Remove old two level fair share thread pool commit_hash:76e62b8b0d0ca4f184947bf2691276f4035d930c
